xiaohei/xiaohei/requirements.txt
httplib2==0.18.1
Django==3.0.7
django_ckeditor==5.9.0
xmltodict==0.12.0
simplejson==3.17.0
Flask==1.1.2
lxml==4.2.5
six==1.12.0
chardet==3.0.4
redis==3.5.3
future==0.18.2
drf_extensions==0.6.0
pymongo==3.10.1
wechatpy==1.8.13
djangorestframework_jwt==1.11.0
xlwt==1.3.0
django_filter==2.2.0
requests==2.23.0
django_crispy_forms==1.9.1
django_filters==0.2.1
django_formtools==2.2
django_reversion==3.0.7
Pillow==7.1.2
pycryptodome==3.9.8
South==1.0.2
xlsxwriter==1.2.9
